Move stark-brainfuck to separate crate in Cargo workspace
BrainSTARK is a proof-of-concept zk-STARK engine built with cryptography
from the twenty-first library. The tutorial for this code is found at:

https://aszepieniec.github.io/stark-brainfuck/

BrainSTARK has lived as a module in twenty-first until it is now being
moved into its own crate under a Cargo workspace. This means BrainSTARK
is continuously compiled, tested and benchmarked against twenty-first,
but is not distributed in the twenty-first crate at crates.io:

https://crates.io/crates/twenty-first
