+++ stark-brainfuck/src/instruction_table.rs
@@ -1,18 +1,17 @@
 use num_traits::{One, Zero};
-
-use crate::shared_math::b_field_element as bfe;
-use crate::shared_math::stark::brainfuck::stark::TERMINAL_COUNT;
-use crate::shared_math::stark::brainfuck::vm::instruction_zerofier;
-use crate::shared_math::{
-    b_field_element::BFieldElement, mpolynomial::MPolynomial, other, x_field_element::XFieldElement,
-};
 use std::convert::TryInto;
 
-use super::{
-    stark::{EXTENSION_CHALLENGE_COUNT, PERMUTATION_ARGUMENTS_COUNT},
-    table::{Table, TableMoreTrait, TableTrait},
-    vm::InstructionMatrixBaseRow,
-};
+use twenty_first::shared_math::b_field_element as bfe;
+use twenty_first::shared_math::b_field_element::BFieldElement;
+use twenty_first::shared_math::mpolynomial::MPolynomial;
+use twenty_first::shared_math::other;
+use twenty_first::shared_math::x_field_element::XFieldElement;
+
+use crate::stark::TERMINAL_COUNT;
+use crate::stark::{EXTENSION_CHALLENGE_COUNT, PERMUTATION_ARGUMENTS_COUNT};
+use crate::table::{Table, TableMoreTrait, TableTrait};
+use crate::vm::instruction_zerofier;
+use crate::vm::InstructionMatrixBaseRow;
 
 #[derive(Debug, Clone)]
 pub struct InstructionTable(pub Table<InstructionTableMore>);
@@ -405,13 +404,15 @@ impl TableTrait for InstructionTable {
 #[cfg(test)]
 mod instruction_table_tests {
     use super::*;
-    use crate::shared_math::stark::brainfuck::vm::sample_programs;
-    use crate::shared_math::stark::brainfuck::vm::BaseMatrices;
-    use crate::shared_math::stark::brainfuck::vm::InstructionMatrixBaseRow;
-    use crate::shared_math::stark::brainfuck::{self};
-    use crate::shared_math::traits::GetRandomElements;
-    use crate::shared_math::traits::PrimitiveRootOfUnity;
     use rand::thread_rng;
+
+    use twenty_first::shared_math::traits::GetRandomElements;
+    use twenty_first::shared_math::traits::PrimitiveRootOfUnity;
+
+    use crate as brainfuck;
+    use crate::vm::sample_programs;
+    use crate::vm::BaseMatrices;
+    use crate::vm::InstructionMatrixBaseRow;
 
     // When we simulate a program, this generates a collection of matrices that contain
     // "abstract" execution traces. When we evaluate the base transition constraints on
