+++ twenty-first/src/shared_math.rs
@@ -8,4 +8,3 @@ pub mod rescue_prime_regular;
 pub mod stark;
 pub mod traits;
 pub mod x_field_element;
-pub mod xfri;
