+++ twenty-first/src/shared_math/stark.rs
@@ -1,3 +1,2 @@
-pub mod brainfuck;
 pub mod rescue_prime;
 pub mod stark_verify_error;
